backend/requirements.txt
fastapi
uvicorn
pymongo
pillow
torch
torchvision
